Build 31012301

+++ ESP32/DIY-Flow-Bench/sensors.cpp
@@ -301,7 +301,7 @@ double Sensors::getMafVolts() {
  *
  * @note MAF decode is done in Calculations.cpp
  ***/
-int Sensors::getMafRaw() {
+long Sensors::getMafRaw() {
 	
 	Hardware _hardware;
 	#ifdef MAF_IS_ENABLED
